third_party/blink/web_tests/platform/generic/fast/forms/ValidityState-tooShort-textarea-expected.txt

Tests for tooShort flag with <textarea> elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

No minlength and no value
PASS textarea.validity.tooShort is false


Non-dirty value
PASS textarea.value.length is 2
PASS textarea.validity.tooShort is false
PASS textarea.value.length is 1
PASS textarea.validity.tooShort is false


Dirty value and longer than minLength
PASS textarea.value.length is 1
PASS textarea.validity.tooShort is true
PASS textarea.value.length is 0
PASS textarea.validity.tooShort is false
PASS textarea.value.length is 2
PASS textarea.validity.tooShort is true
PASS textarea.value.length is 3
PASS textarea.validity.tooShort is false


Sets a value via DOM property
PASS textarea.validity.tooShort is false


Disabling makes the control valid
PASS textarea.validity.tooShort is true
PASS textarea.disabled = true; textarea.validity.tooShort is false
PASS textarea.disabled = false; textarea.validity.tooShort is true


A value set by resetting a form doesn't make tooShort true.
PASS textarea.validity.tooShort is true
PASS textarea.value is "ab"
PASS textarea.validity.tooShort is false


A value set by a child node change doesn't make tooShort true.
PASS textarea.validity.tooShort is false
PASS textarea.value is "a"
PASS textarea.validity.tooShort is false


Grapheme length is shorter than minLength though character length is greater
PASS textarea.validity.tooShort is false
PASS textarea.value.length is 3
PASS textarea.validity.tooShort is false
PASS successfullyParsed is true

TEST COMPLETE
